requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0

# ── Vision Transformer backbone ───────────────────────────────────────────
timm>=0.9.0

# ── Diffusion model (optional — needed for real facial generation) ────────
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0

# ── Image processing ──────────────────────────────────────────────────────
Pillow>=9.5.0
opencv-python>=4.8.0

# ── Evaluation & statistics ───────────────────────────────────────────────
scikit-learn>=1.3.0

# ── Training utilities ───────────────────────────────────────────────────
tqdm>=4.65.0

# ── Notebook / visualisation (optional) ──────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
